+++ docs/history.md
@@ -5,11 +5,13 @@
 
 ### _v1.17.0_ - September 17th, 2023 (Unreleased)
 
+- Added the following babies:
+  - Multidimensional Baby (#593) - Spawns 3 Swarm Spiders on death
 - The following babies have new effects:
+  - Hopeless Baby (#125) - "+2 keys + keys are hearts" --> "Keys are hearts"
+  - Mohawk Baby (#138) - "+2 bombs + bombs are hearts" --> "Bombs are hearts"
   - Spiky Demon Baby (#277) - "Starts with Dark Arts" --> "Starts with Dark Arts + Car Battery"
   - Cupcake Baby (#321) - "Starts with The Intruder" --> "Starts with Tear Detonator + 9 Volt + Car Battery"
-- coin health
-- spawns Swarm Spiders on death
 
 ### _v1.16.1_ - September 16th, 2023